fixed testcases, and did proper contexting of entity manager
