--- apps/acqdat_api/lib/acqdat_api/asset_type.ex
@@ -1,6 +1,0 @@
-defmodule AcqdatApi.AssetType do
-  alias AcqdatCore.Model.AssetType, as: AssetTypeModel
-
-  defdelegate get(id), to: AssetTypeModel
-  defdelegate update_asset(asset_type, data), to: AssetTypeModel
-end
