+++ apps/acqdat_api/lib/acqdat_api/entity-management/asset_type.ex
@@ -1,0 +1,6 @@
+defmodule AcqdatApi.EntityManagement.AssetType do
+  alias AcqdatCore.Model.EntityManagement.AssetType, as: AssetTypeModel
+
+  defdelegate get(id), to: AssetTypeModel
+  defdelegate update_asset(asset_type, data), to: AssetTypeModel
+end
