+++ apps/acqdat_api/lib/acqdat_api_web/controllers/entity-management/asset_controller.ex
@@ -1,6 +1,6 @@
-defmodule AcqdatApiWeb.AssetController do
+defmodule AcqdatApiWeb.EntityManagement.AssetController do
   use AcqdatApiWeb, :controller
-  alias AcqdatApi.Asset
+  alias AcqdatApi.EntityManagement.Asset
   import AcqdatApiWeb.Helpers
 
   plug :load_asset when action in [:show, :update]
