+++ apps/acqdat_api/lib/acqdat_api_web/controllers/entity-management/entity_controller.ex
@@ -1,6 +1,6 @@
-defmodule AcqdatApiWeb.EntityController do
+defmodule AcqdatApiWeb.EntityManagement.EntityController do
   use AcqdatApiWeb, :controller
-  alias AcqdatApi.EntityParser
+  alias AcqdatApi.EntityManagement.EntityParser
   alias AcqdatCore.Model.EntityManagement.Organisation, as: OrgModel
   import AcqdatApiWeb.Helpers
