+++ apps/acqdat_api/lib/acqdat_api_web/views/entity-management/asset_type_view.ex
@@ -1,4 +1,4 @@
-defmodule AcqdatApiWeb.AssetTypeView do
+defmodule AcqdatApiWeb.EntityManagement.AssetTypeView do
   use AcqdatApiWeb, :view
 
   def render("asset_type.json", %{asset_type: asset_type}) do
