+++ apps/acqdat_api/lib/acqdat_api_web/views/entity-management/entity_view.ex
@@ -1,6 +1,6 @@
-defmodule AcqdatApiWeb.EntityView do
+defmodule AcqdatApiWeb.EntityManagement.EntityView do
   use AcqdatApiWeb, :view
-  alias AcqdatApiWeb.AssetView
+  alias AcqdatApiWeb.EntityManagement.AssetView
 
   def render("organisation_tree.json", %{org: org}) do
     %{
