+++ apps/acqdat_api/test/acqdat_api_web/controllers/role-management/team_controller_test.exs
@@ -129,8 +129,7 @@ defmodule AcqdatApiWeb.RoleManagement.TeamControllerTest do
 
     setup do
       org = insert(:organisation)
-      project = insert(:project)
-      asset = insert(:asset, project_id: project.id)
+      asset = insert(:asset)
       team = insert(:team)
 
       [team: team, asset: asset, org: org]
