+++ apps/acqdat_api/test/acqdat_api_web/controllers/role-management/user_controller_test.exs
@@ -38,8 +38,7 @@ defmodule AcqdatApiWeb.RoleManagement.UserControllerTest do
 
     setup do
       org = insert(:organisation)
-      project = insert(:project)
-      asset = insert(:asset, project_id: project.id)
+      asset = insert(:asset)
       user = insert(:user)
 
       [user: user, asset: asset, org: org]
