+++ apps/acqdat_core/test/role-management/schema/user_asset_test.exs
@@ -9,8 +9,7 @@ defmodule AcqdatCore.Schema.RoleManagement.UserAssetTest do
   describe "changeset/2" do
     setup do
       user = insert(:user)
-      project = insert(:project)
-      asset = insert(:asset, project_id: project.id)
+      asset = insert(:asset)
 
       [user: user, asset: asset]
     end
